backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.23
pydantic==2.1.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.7
supabase==2.1.0
httpx==0.27.0
requests==2.31.0
razorpay==1.4.1
googlemaps==4.10.0
pytest==7.4.0
alembic==1.13.1
psycopg2-binary==2.9.9 